backend/atomicMicroservices/booking/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
mysql-connector-python==9.2.0
SQLAlchemy==2.0.39
flask-cors==5.0.1
python-dotenv==1.0.1
